chore: Move clusterresourceset helpers to the correct file
