+++ exp/addons/internal/controllers/clusterresourceset_controller.go
@@ -18,9 +18,7 @@ package controllers
 
 import (
 	"context"
-	"encoding/base64"
 	"fmt"
-	"sort"
 	"time"
 
 	"github.com/pkg/errors"
@@ -495,47 +493,4 @@ func (r *ClusterResourceSetReconciler) resourceToClusterResourceSet(o client.Obj
 	}
 
 	return result
-}
-
-func getDataListAndHash(resourceKind string, unstructuredData map[string]interface{}, errList []error) ([][]byte, string) {
-	// Since maps are not ordered, we need to order them to get the same hash at each reconcile.
-	keys := make([]string, 0)
-
-	for key := range unstructuredData {
-		keys = append(keys, key)
-	}
-	sort.Strings(keys)
-
-	dataList := make([][]byte, 0)
-	for _, key := range keys {
-		val, ok, err := unstructured.NestedString(unstructuredData, key)
-		if !ok || err != nil {
-			errList = append(errList, errors.New("failed to get value field from the resource"))
-			continue
-		}
-
-		byteArr := []byte(val)
-		// If the resource is a Secret, data needs to be decoded.
-		if resourceKind == string(addonsv1.SecretClusterResourceSetResourceKind) {
-			byteArr, _ = base64.StdEncoding.DecodeString(val)
-		}
-
-		dataList = append(dataList, byteArr)
-	}
-
-	return dataList, computeHash(dataList)
-}
-
-func handleGetResourceErrors(clusterResourceSet *addonsv1.ClusterResourceSet, err error, errList []error) {
-	if err == ErrSecretTypeNotSupported {
-		conditions.MarkFalse(clusterResourceSet, addonsv1.ResourcesAppliedCondition, addonsv1.WrongSecretTypeReason, clusterv1.ConditionSeverityWarning, err.Error())
-	} else {
-		conditions.MarkFalse(clusterResourceSet, addonsv1.ResourcesAppliedCondition, addonsv1.RetrievingResourceFailedReason, clusterv1.ConditionSeverityWarning, err.Error())
-
-		// Continue without adding the error to the aggregate if we can't find the resource.
-		if apierrors.IsNotFound(err) {
-			return
-		}
-	}
-	errList = append(errList, err)
 }
